move GetEdgesFromWeightedGraph test into typesystem_test
